루틴 저장 post_no 입력 안 받아도 되게 수정

+++ src/main/java/com/godMorning_backend/domain/ToDo.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Setter
 public class ToDo {
 
-    private int post_no;
+    private Long post_no;
     private String content;
     private int check_do;
